bot now retains spoiler status of messages sent

+++ main.py
@@ -61,11 +61,17 @@ class Bot(discord.Client):
 
             filter_nonlinks = filter_nonlinks.strip()
 
+            if "||" in message.content:
+                try:
+                    await channel.send(content=f'From: <@{message.author.id}> ||{filter_nonlinks} {filter_links}||')
+                    return await message.delete()
+                except:
+                    return await channel.send(content=f'error sending fixed twitter link')
             try:
                 await channel.send(content=f'From: <@{message.author.id}> {filter_nonlinks} {filter_links}')
-                await message.delete()
+                return await message.delete()
             except:
-                await channel.send(content=f'error sending fixed twitter link')
+                return await channel.send(content=f'error sending fixed twitter link')
 
         if message.content.startswith("!natetimer"):
             try:
